* Updated terminal reaches parser to accept name and id as attributes * Changed id to Long from Integer
Former-commit-id: e260c8ce46d0c5fee10cba6acc99ef7726f977d1

+++ src/test/java/unit/gov/usgswim/sparrow/parser/TerminalReachesTest.java
@@ -33,11 +33,11 @@ public class TerminalReachesTest extends TestCase {
 
 		TerminalReaches termReaches = buildTestInstance(1L);
 		
-		List<Integer> reachIDs = termReaches.getReachIDs();
+		List<Long> reachIDs = termReaches.getReachIDs();
 		assertEquals(3, reachIDs.size());
-		assertEquals(Integer.valueOf(2345642), reachIDs.get(0));
-		assertEquals(Integer.valueOf(3425688), reachIDs.get(1));
-		assertEquals(Integer.valueOf(5235424), reachIDs.get(2));
+		assertEquals(Long.valueOf(2345642), reachIDs.get(0));
+		assertEquals(Long.valueOf(3425688), reachIDs.get(1));
+		assertEquals(Long.valueOf(5235424), reachIDs.get(2));
 
 	}
